Update kode terbaru 09/06/2024

+++ database/migrations/2024_05_28_132232_add_user_id_to_pesertas_table.php
@@ -12,7 +12,9 @@ class AddUserIdToPesertasTable extends Migration
     public function up(): void
     {
         Schema::table('pesertas', function (Blueprint $table) {
-            $table->foreignId('user_id')->constrained()->onDelete('cascade')->after('id');
+            if (!Schema::hasColumn('pesertas', 'user_id')) {
+                $table->foreignId('user_id')->constrained()->onDelete('cascade')->after('id');
+            }
         });
     }
